popm: remove support for WebAssembly
We no longer plan to support WebAssembly in the PoP Miner. As such, this
removes the @hemilabs/pop-miner NPM package (API wrapper), the WASM PoP
miner itself, and supporting APIs in the service/popmd package.
