+++ .github/CODEOWNERS
@@ -30,9 +30,6 @@
 /database/tbcd/ @hemilabs/dev @marcopeereboom
 /service/tbc/   @hemilabs/dev @marcopeereboom
 
-# WASM PoP Miner (popmd/wasm)
-/web/ @hemilabs/dev @joshuasing
-
 # Release script
 /scripts/release.sh @hemilabs/dev @joshuasing
