+++ .github/workflows/release.yml
@@ -26,105 +26,6 @@ jobs:
     name: "Test (Go)"
     uses: ./.github/workflows/go.yml
 
-  # Prepare to release
-  prepare:
-    name: "Prepare"
-    runs-on: "ubuntu-latest"
-    permissions:
-      contents: read
-    outputs:
-      version: "${{ steps.version.outputs.version }}"
-      tag: "${{ steps.version.outputs.tag }}"
-      version_type: "${{ steps.version.outputs.type }}"
-    steps:
-      - name: "Determine version type"
-        id: version
-        env:
-          RAW_VERSION: "${{ github.ref_name }}"
-        # This script determines the version type (stability), e.g.
-        #  1.0.0 = stable, 1.1.0-rc.1 = unstable, 0.1.0 = unstable
-        run: |
-          VERSION=$(echo "$RAW_VERSION" | sed -e 's/^v//')
-          TAG=$(echo "$RAW_VERSION" | sed -E 's/^([^v])/v\1/g')
-          echo "version=$VERSION" >> "$GITHUB_OUTPUT"
-          echo "tag=$TAG" >> "$GITHUB_OUTPUT"
-
-          TYPE=unstable
-          if echo "$VERSION" | grep -Eq '^[1-9][0-9]*\.[0-9]+\.[0-9]+$'; then
-            TYPE=stable
-          fi
-          echo "Detected that $TAG is $TYPE"
-          echo "type=$TYPE" >> "$GITHUB_OUTPUT"
-
-  # Publish NPM packages
-  npm:
-    name: "npm"
-    runs-on: "ubuntu-latest"
-    needs: [ "prepare" ]
-    permissions:
-      contents: read
-    steps:
-      - name: "Checkout repository"
-        uses: actions/checkout@11bd71901bbe5b1630ceea73d27597364c9af683 # v4.2.2
-
-      - name: "Setup Go ${{ env.GO_VERSION }}"
-        uses: actions/setup-go@41dfa10bad2bb2ae585af6ee5bb4d7d973ad74ed # v5.1.0
-        with:
-          go-version: "${{ env.GO_VERSION }}"
-          cache: true
-          check-latest: true
-
-      - name: "Setup pnpm ${{ env.PNPM_VERSION }}"
-        uses: pnpm/action-setup@0c17529a66aca453f9227af23103ed11469b1e47 # v4.0.0
-        with:
-          version: "${{ env.PNPM_VERSION }}"
-
-      - name: "Setup Node"
-        uses: actions/setup-node@39370e3970a6d050c480ffad4ff0ed4d3fdee5af # v4.1.0
-        with:
-          node-version-file: "web/.nvmrc"
-          check-latest: true
-          cache: "pnpm"
-          cache-dependency-path: "web/**/pnpm-lock.yaml"
-          registry-url: "https://registry.npmjs.org"
-
-      - name: "Install dependencies"
-        working-directory: "web/"
-        run: pnpm install --frozen-lockfile
-
-      - name: "Install Go dependencies"
-        working-directory: "web/"
-        run: make deps
-
-      - name: "Set package.json versions"
-        working-directory: "web/"
-        env:
-          VERSION: "${{ needs.prepare.outputs.version }}"
-        run: |
-          # Prints all package.json files
-          PACKAGE_FILES=$(find . -path '**/node_modules' -prune -o -name 'package.json' -print)
-          for file in $PACKAGE_FILES; do
-            # Set "version" in package.json file to $VERSION.
-            TMP_FILE="$(mktemp)"
-            jq --arg v "$VERSION" '.version = $v' "$file" > "$TMP_FILE"
-            mv "$TMP_FILE" "$file"
-          done
-
-        # TODO(joshuasing): Install and use binaryen
-      - name: "Build @hemilabs/pop-miner WebAssembly binary"
-        working-directory: "web/"
-        run: make wasm
-
-      - name: "Build @hemilabs/pop-miner package"
-        working-directory: "web/"
-        run: pnpm build:pop-miner
-
-      - name: "Publish @hemilabs/pop-miner package"
-        working-directory: "web/"
-        run: pnpm publish packages/pop-miner --access public --no-git-checks
-        env:
-          NODE_AUTH_TOKEN: "${{ secrets.NPM_TOKEN }}"
-
   # Create GitHub Release
   release:
     name: "Release"
