--- service/popm/prometheus_wasm.go
@@ -1,13 +1,0 @@
-// Copyright (c) 2024 Hemi Labs, Inc.
-// Use of this source code is governed by the MIT License,
-// which can be found in the LICENSE file.
-
-//go:build js && wasm
-
-package popm
-
-import "context"
-
-func (m *Miner) handlePrometheus(ctx context.Context) error {
-	return nil
-}
