--- web/.gitignore
@@ -1,4 +1,0 @@
-webapp/
-node_modules/
-dist/
-packages/pop-miner/wasm/
