--- web/.nvmrc
@@ -1,1 +1,0 @@
-20.15.0
