--- web/.prettierignore
@@ -1,5 +1,0 @@
-**/dist
-pnpm-lock.yaml
-
-www/
-webapp/
