--- web/package.json
@@ -1,40 +1,0 @@
-{
-  "name": "@hemilabs/heminetwork",
-  "version": "0.0.0",
-  "private": true,
-  "license": "MIT",
-  "bugs": "https://github.com/hemilabs/heminetwork/issues",
-  "homepage": "https://github.com/hemilabs/heminetwork",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/hemilabs/heminetwork.git",
-    "directory": "web"
-  },
-  "scripts": {
-    "preinstall": "npx only-allow pnpm",
-    "build": "make wasm",
-    "build:opt": "make wasm-opt",
-    "build:pop-miner": "pnpm --filter @hemilabs/pop-miner run build",
-    "lint": "eslint \"packages/**/src/**/*.ts\" && prettier . --check",
-    "format": "prettier -w ."
-  },
-  "packageManager": "pnpm@9.4.0",
-  "workspaces": [
-    "packages/*"
-  ],
-  "engines": {
-    "node": "^18.0.0 || >=20.0.0"
-  },
-  "devDependencies": {
-    "@eslint/js": "9.13.0",
-    "eslint": "9.13.0",
-    "eslint-config-prettier": "9.1.0",
-    "eslint-plugin-prettier": "5.2.1",
-    "eslint-plugin-tsdoc": "0.3.0",
-    "globals": "15.11.0",
-    "prettier": "3.3.3",
-    "typescript": "5.6.3",
-    "typescript-eslint": "8.11.0",
-    "vitest": "2.1.3"
-  }
-}
