--- web/packages/pop-miner/package.json
@@ -1,43 +1,0 @@
-{
-  "name": "@hemilabs/pop-miner",
-  "version": "0.1.0",
-  "type": "module",
-  "license": "MIT",
-  "author": "Hemi Labs",
-  "description": "Hemi Labs WebAssembly Proof-of-Proof Miner",
-  "repository": {
-    "type": "git",
-    "url": "https://github.com/hemilabs/heminetwork.git",
-    "directory": "web/packages/pop-miner"
-  },
-  "homepage": "https://hemi.xyz",
-  "bugs": {
-    "url": "https://github.com/hemilabs/heminetwork/issues"
-  },
-  "engines": {
-    "node": "^18.0.0 || >=20.0.0"
-  },
-  "scripts": {
-    "build": "tsup"
-  },
-  "main": "./dist/browser/index.js",
-  "types": "./dist/types.d.ts",
-  "files": [
-    "dist"
-  ],
-  "exports": {
-    ".": {
-      "types": "./dist/browser/index.d.ts",
-      "browser": "./dist/browser/index.js",
-      "import": "./dist/browser/index.js",
-      "default": "./dist/browser/index.js"
-    },
-    "./popminer.wasm": "./dist/popminer.wasm",
-    "./types": "./dist/types.d.ts",
-    "./package.json": "./package.json"
-  },
-  "devDependencies": {
-    "tsup": "8.3.4",
-    "typescript": "5.6.3"
-  }
-}
