--- web/www/index.html
@@ -1,144 +1,0 @@
-<!--
- Copyright (c) 2024 Hemi Labs, Inc.
- Use of this source code is governed by the MIT License,
- which can be found in the LICENSE file.
--->
-
-<html>
-  <head>
-    <meta charset="UTF-8" />
-    <title>PoP miner integration tests</title>
-    <script src="wasm_exec.js"></script>
-    <script src="popminer.js"></script>
-  </head>
-  <body>
-    <div>
-      <button id="VersionButton">version</button>
-      <pre class="VersionShow"></pre>
-    </div>
-
-    <div>
-      <input id="GenerateKeyNetworkInput" value="testnet3">
-      <label for="GenerateKeyNetworkInput">bitcoin network</label>
-      <br>
-
-      <button id="GenerateKeyButton">generate key</button>
-      <pre class="GenerateKeyShow"></pre>
-    </div>
-
-    <div>
-      <input id="ParseKeyNetworkInput" value="testnet3">
-      <label for="ParseKeyNetworkInput">bitcoin network</label>
-      <br>
-      <input id="ParseKeyPrivateKeyInput" type="password">
-      <label for="ParseKeyPrivateKeyInput">secp256k1 private key (hexadecimal string, 32 bytes)</label>
-      <br>
-
-      <button id="ParseKeyButton">parse key</button>
-      <pre class="ParseKeyShow"></pre>
-    </div>
-
-    <div>
-      <input id="BitcoinAddressToScriptHashNetworkInput" value="testnet3">
-      <label for="BitcoinAddressToScriptHashNetworkInput">bitcoin network</label>
-      <br>
-      <input id="BitcoinAddressToScriptHashAddressInput">
-      <label for="BitcoinAddressToScriptHashAddressInput">bitcoin address</label>
-      <br>
-
-      <button id="BitcoinAddressToScriptHashAddressButton">bitcoin address to script hash</button>
-      <pre class="BitcoinAddressToScriptHashAddressShow"></pre>
-    </div>
-
-    <div>
-      <input id="StartPopMinerLogLevelInput" value="hemiwasm=INFO:popm=INFO:protocol=INFO">
-      <label for="StartPopMinerLogLevelInput">log level</label>
-      <br>
-
-      <input id="StartPopMinerNetworkInput" value="testnet">
-      <label for="StartPopMinerNetworkInput">Hemi network (testnet, devnet or mainnet)</label>
-      <br>
-
-      <input id="StartPopMinerPrivateKeyInput" type="password">
-      <label for="StartPopMinerPrivateKeyInput">bitcoin private key (must be funded)</label>
-      <br>
-      <input id="StartPopMinerAutomaticFeesInput" value="fastest">
-      <label for="StartPopMinerAutomaticFeesInput">
-          Whether to do automatic fees (set to 'false' to disable), and which type of recommended fee to use.
-      </label>
-      <br>
-      <input id="StartPopMinerAutomaticFeeMultiplierInput" value="1.1">
-      <label for="StartPopMinerAutomaticFeeMultiplierInput">Fee multiplier</label>
-      <br>
-      <input id="StartPopMinerAutomaticFeeRefreshInput" value="300">
-      <label for="StartPopMinerAutomaticFeeRefreshInput">Interval to refresh recommended fees from mempool.space API</label>
-      <br>
-      <input id="StartPopMinerStaticFeeInput" value="50">
-      <label for="StartPopMinerStaticFeeInput">Fee to pay for PoP Mining (sat/vB)</label>
-      <br>
-
-      <button id="StartPopMinerButton">Run PoP Miner</button>
-      <pre class="StartPopMinerShow"></pre>
-    </div>
-
-    <div>
-      <button id="StopPopMinerButton">Stop PoP Miner</button>
-      <pre class="StopPopMinerShow"></pre>
-    </div>
-
-    <div>
-      <button id="minerStatusButton">Miner Status</button>
-      <pre class="minerStatusDisplay"></pre>
-    </div>
-
-    <!-- connected commands go here, run after StartPopMinerButton -->
-    <hr>
-    <p>The following commands require a live connection (after StartPopMinerButton)</p>
-    <br>
-
-    <div>
-      <button id="PingButton">ping</button>
-      <pre class="PingShow"></pre>
-    </div>
-
-    <div>
-      <input id="L2KeystonesNumL2KeystonesInput" value="2" type="number">
-      <label for="L2KeystonesNumL2KeystonesInput">num l2 keystones</label>
-      <br>
-
-      <button id="L2KeystonesButton">l2 keystones</button>
-      <pre class="L2KeystonesShow"></pre>
-    </div>
-
-    <div>
-      <input id="BitcoinBalanceScriptHashInput" value="xxx">
-      <label for="BitcoinBalanceScriptHashInput">script hash</label>
-      <br>
-
-      <button id="BitcoinBalanceButton">bitcoin balance</button>
-      <pre class="BitcoinBalanceShow"></pre>
-    </div>
-
-    <div>
-      <button id="BitcoinInfoButton">bitcoin info</button>
-      <pre class="BitcoinInfoShow"></pre>
-    </div>
-
-    <div>
-      <input id="BitcoinUTXOsScriptHashInput" value="xxx">
-      <label for="BitcoinUTXOsScriptHashInput">script hash</label>
-      <br>
-
-      <button id="BitcoinUTXOsButton">bitcoin utxos</button>
-      <pre class="BitcoinUTXOsShow"></pre>
-    </div>
-
-    <div>
-      <br><hr><br>
-      <p>Events</p>
-      <pre style="height: 50vh; text-wrap: normal; overflow-y: scroll;" class="eventsDisplay"></pre>
-    </div>
-
-    <script src="index.js" defer></script>
-  </body>
-</html>
